requirements.txt
google-play-scraper
pandas
ipykernel
transformers 
torch
scikit-learn
spacy
psycopg2-binary
wordcloud
pytest

seaborn
matplotlib